requirements.txt
tkinter
plantuml
Pillow